requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.9.1
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
ccxt==1.74.11
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
commonmark==0.9.1
cryptography==36.0.1
frozenlist==1.3.0
idna==3.3
iniconfig==1.1.1
multidict==6.0.2
numpy==1.22.2
packaging==21.3
pandas==1.4.1
pandas-ta==0.3.14b0
pluggy==1.0.0
py==1.11.0
pyaes==1.6.1
pyasn1==0.4.8
pycares==4.1.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rich==11.2.0
rsa==4.8
six==1.16.0
style==1.1.0
Telethon==1.24.0
tomli==2.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
yarl==1.7.2
